Add docs for grid layout
