+++ vgtk/src/lib.rs
@@ -70,7 +70,6 @@
 mod callback;
 mod component;
 pub mod ext;
-pub mod grid;
 mod menu_builder;
 #[doc(hidden)]
 pub mod properties;
